ENH: ELASTIX-9 Only access connections via component indexes; BUG: ELASTIX-9 ComponentExist() did not work
The ENH: We deliberately avoid using connection indexes, but instead
force the user to think in terms of components which is conceptually
simpler and results in less pointers being thrown around while
constructing the blueprint.

The BUG: Temporarily disable functionality for deleting component until
we find a god way of asserting that vertex exist before it is accessed.