STYLE: Replaced BOOST_FOREACH by `if`, in `BlueprintImpl::FindIncludes`
I think we should not use BOOST_FOREACH anymore, as it is basically superseded by C++11 range-based for-loop. However, in `BlueprintImpl::FindIncludes`, could not be replaced by a range-based for-loop, as it used an std::pair of iterators. But in this specific case, an `if` statement will do  :-)

Fixes VS2017 /analyze warning C26477 "Use 'nullptr' rather than 0 or NULL (es.47)", which came from BOOST_FOREACH.

Also simplified the lines of code that convert strings to paths.